refactor api extractor class, move all logic into subclasses from extract command
